+++ opnieuw/util.py
@@ -1,14 +1,13 @@
 from __future__ import annotations
 
-from collections.abc import Iterator
 from contextlib import AbstractContextManager
 
-from .retries import Action, DoCall, RetryState, replace_retry_state
+from .retries import BackoffCalculator, replace_backoff_calculator
 
 
-class NoRetryState(RetryState):
-    def __iter__(self) -> Iterator[Action]:
-        yield DoCall()
+class NoRetryBackoff(BackoffCalculator):
+    def get_backoff(self) -> float | None:
+        return None
 
 
 def no_retries(namespace: str | None = None) -> AbstractContextManager[None]:
@@ -17,4 +16,4 @@ def no_retries(namespace: str | None = None) -> AbstractContextManager[None]:
     `retry_async` decorators with the provided namespace. None means all decorators
     without a provided namespace will not retry.
     """
-    return replace_retry_state(NoRetryState, namespace=namespace)
+    return replace_backoff_calculator(NoRetryBackoff, namespace=namespace)